Mods for testing one cost function at a time; something is wrong with acceleration handling - seems "stepwise" when it should be continuous - but lane changing does occur with the single cost function

+++ Term 3/06 - Project - Path Planning/src/vehicle.h
@@ -62,7 +62,7 @@ class Vehicle {
 
   string get_next_state(map<int, vector <vector<int>>> predictions);
 
-  string min_cost_state(vector<string> states, vector<int> costs);
+  string min_cost_state(vector<string> states, vector<double> costs);
 
   vector<Snapshot> trajectories_for_state(string state, map<int, vector <vector<int>>> predictions, int horizon = 5);
